requirements.txt
asgiref==3.10.0
Django==4.2.25
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
